Adding new dlsym() method for loading TAU The code now supports weak/strong symbols and dlopen/dlsym methods for loading TAU symbols.  On some platforms, when built as part of HPX, TAU symbols aren't overridden when built with the clang compiler.  The new default method is to use dlsym() to load the symbols, and the old weak/strong symbol method can be used if the APEX_USE_WEAK_SYMBOLS CMake option is set to TRUE or ON.

+++ src/unit_tests/C++/apex_non_worker_thread.cpp
@@ -3,6 +3,7 @@
 #include <stdio.h>
 #include <thread>
 #include <string>
+#include <cinttypes>
 #include "thread_instance.hpp"
 
 #define MAX_OUTER 50
@@ -11,7 +12,7 @@
 
 uint64_t func(uint64_t i) {
     char name[128];
-    sprintf(name, "func %lu", i);
+    sprintf(name, "func %llu", static_cast<unsigned long long>(i));
     apex::profiler* p = apex::start(std::string(name));
     uint64_t j = i * i;
     apex::stop(p);
